LPS-199140 refactor SharingHelper.java change to a util class and pass only reference by method usage on getSharingEntryPermissionDisplays, and apply to usages passing new reference
