+++ modules/apps/sharing/sharing-web/src/main/java/com/liferay/sharing/web/internal/util/SharingHelperUtil.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.sharing.web.internal.helper;
+package com.liferay.sharing.web.internal.util;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
@@ -22,16 +22,12 @@ import java.util.List;
 import java.util.Locale;
 import java.util.ResourceBundle;
 
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
 /**
  * @author Sergio González
  */
-@Component(service = SharingHelper.class)
-public class SharingHelper {
+public class SharingHelperUtil {
 
-	public SharingEntryPermissionDisplayAction
+	public static SharingEntryPermissionDisplayAction
 		getSharingEntryPermissionDisplayActionKey(SharingEntry sharingEntry) {
 
 		if (sharingEntry.hasSharingPermission(SharingEntryAction.UPDATE)) {
@@ -51,9 +47,10 @@ public class SharingHelper {
 		return null;
 	}
 
-	public List<SharingEntryPermissionDisplay>
+	public static List<SharingEntryPermissionDisplay>
 		getSharingEntryPermissionDisplays(
-			PermissionChecker permissionChecker, long classNameId, long classPK,
+			PermissionChecker permissionChecker,
+			SharingPermission sharingPermission, long classNameId, long classPK,
 			long groupId, Locale locale) {
 
 		List<SharingEntryAction> sharingEntryActions = new ArrayList<>();
@@ -62,7 +59,7 @@ public class SharingHelper {
 				SharingEntryAction.values()) {
 
 			try {
-				if (_sharingPermission.contains(
+				if (sharingPermission.contains(
 						permissionChecker, classNameId, classPK, groupId,
 						Collections.singletonList(sharingEntryAction))) {
 
@@ -75,15 +72,13 @@ public class SharingHelper {
 		}
 
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
-			locale, SharingHelper.class);
+			locale, SharingHelperUtil.class);
 
 		return SharingEntryPermissionDisplay.getSharingEntryPermissionDisplays(
 			sharingEntryActions, resourceBundle);
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(SharingHelper.class);
-
-	@Reference
-	private SharingPermission _sharingPermission;
+	private static final Log _log = LogFactoryUtil.getLog(
+		SharingHelperUtil.class);
 
 }
